Loader component renamed as TreeLoader, Both Treeloader & API Loader moved to Loaders folder in Utils folder.
